requirements.txt
streamlit
google-generativeai
pyperclip